requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
beautifulsoup4>=4.12.0
requests
lxml
vaderSentiment
spacy>=3.7.0
apscheduler
python-dotenv
openai